+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.13.1
+ * Update CLDR data to version 29.
 
 ## 0.13.0
  * Add support for compact number formats ("1.2K") and for significant digits in